refactored dashboard component
